refactor(storage): remove deprecated namespaced API
